Add primary or spouse attr reader to state file person
